+++ src/Awesomite/StackTrace/VarDumpers/Properties/Property.php
@@ -7,7 +7,7 @@ use Awesomite\StackTrace\Exceptions\InvalidArgumentException;
 /**
  * @internal
  */
-class ReflectionProperty implements PropertyInterface
+class Property implements PropertyInterface
 {
     private $reflection;
 
@@ -22,31 +22,6 @@ class ReflectionProperty implements PropertyInterface
         $this->object = $object;
     }
 
-    public function getName()
-    {
-        return $this->reflection->getName();
-    }
-
-    public function isPrivate()
-    {
-        return $this->reflection->isPrivate();
-    }
-
-    public function isProtected()
-    {
-        return $this->reflection->isProtected();
-    }
-
-    public function isPublic()
-    {
-        return $this->reflection->isPublic();
-    }
-
-    public function isStatic()
-    {
-        return $this->reflection->isStatic();
-    }
-
     public function getValue()
     {
         $this->reflection->setAccessible(true);
@@ -55,8 +30,8 @@ class ReflectionProperty implements PropertyInterface
         return $result;
     }
 
-    public function getDeclaringClass()
+    public function getReflection()
     {
-        return $this->reflection->getDeclaringClass()->getName();
+        return $this->reflection;
     }
 }