+++ src/Awesomite/StackTrace/VarDumpers/Properties/PropertyInterface.php
@@ -7,35 +7,10 @@ namespace Awesomite\StackTrace\VarDumpers\Properties;
  */
 interface PropertyInterface
 {
-    /**
-     * @return string
-     */
-    public function getName();
-
-    /**
-     * @return bool
-     */
-    public function isStatic();
-
-    /**
-     * @return bool
-     */
-    public function isPublic();
-
-    /**
-     * @return bool
-     */
-    public function isProtected();
-
-    /**
-     * @return bool
-     */
-    public function isPrivate();
-
     public function getValue();
 
     /**
-     * @return string
+     * @return \ReflectionProperty
      */
-    public function getDeclaringClass();
+    public function getReflection();
 }